### 2021-06-25 #3409 @kyleniemeyer
Hi @VBlesius, thanks for your interest in JOSS!

Unfortunately, we have extremely limited editor availability to handle submissions right now, and so we need to put your submission on our waitlist untill an appropriate editor becomes available. Thanks for your patience!

Also, I think you need to change your paper and ref filenames to `paper.md` and `paper.bib` for our bot to find it.
