### 2021-06-25 8de6665
adds comment regarding normalisation order to vignette
