### 2021-06-25 0816025
changes dummydata creation and all files related on testdata
Before, the VPCS of dummy data included 6 intervals with an identical
slope. This lead to round-off errors in the CRAN tests "without long
double" and this to switching values of TT between 3 and 4.

Changed the HR decrease part to 5 intervals, so that TT is clearly 3.
