### 2021-06-28 e1d6581
renames bibliography file for the JOSS bot to find it
