### 2021-06-28 #3409 @VBlesius
Unfortunately there don't seem to be reviewers that use R with a focus on health informatics / cardiology. The reviewers with the closest research interests could be:
    - shahmoradi
    - Path2Genome
    - carleshf
    - sachsmc
    - rrrlw
    - Suresh-lab
    - Gweissman
    - darokun
    - stensy
    - acolum
    - Rmadillo
    - Jminnier
    - pacoramon
    - peterakirk
