### 2021-07-15 #3409 @kthyng
Hi @bmcfee! Are you interested in taking on this submission? I thought you may have relevant expertise from the signal processing side.
