### 2021-07-16 #3409 @bmcfee
@VBlesius :wave: I'm tracking down editors now.  Meanwhile, I've noticed that the repo doesn't have explicit license file (see [review checklist](https://joss.readthedocs.io/en/latest/review_checklist.html#general-checks)), though GPLv2 is noted in the readme.  Otherwise, everything seems to be in good shape.
