### 2021-07-19 #3409 @VBlesius
Hi @bmcfee !

> I'm tracking down reviewers now.
Thanks a lot!

> Meanwhile, I've noticed that the repo doesn't have explicit license file (see [review checklist](https://joss.readthedocs.io/en/latest/review_checklist.html#general-checks)), though GPLv2 is noted in the readme.
Right, thank you for the hint! I have added a licence file to the repo.
