### 2021-07-26 #3540 @whedon
[REVIEW]: RHRT: An R package to assess Heart Rate Turbulence
**Submitting author:** @VBlesius (<a href="http://orcid.org/0000-0002-2391-242X">Valeria Blesius</a>)
**Repository:** <a href="https://github.com/VBlesius/RHRT" target ="_blank">https://github.com/VBlesius/RHRT</a>
**Version:** 1.0
