### 2021-08-01 VBlesius/RHRT#3 @rrrlw
Incomplete LICENSE file
[LICENSE](https://github.com/VBlesius/RHRT/blob/main/LICENSE) file (GPL-2) is missing a couple of fields:
* line 293
* line 294
* could adjust the sample (line 312 to end)

Part of [this JOSS review](https://github.com/openjournals/joss-reviews/issues/3540).
