### 2021-08-01 #3540 @rrrlw
@VBlesius: the paper lists 2 authors, yourself and Andreas Dominik. I only see commits from you in the GitHub repository. Could you provide (brief) insight regarding the contributions of the second author?
