### 2021-08-01 VBlesius/RHRT#4 @rrrlw
Contributor guidelines
I don't see any explicit contributor guidelines in this repository. If I missed them, please share the link and I will correct my oversight. Otherwise, options for reasonable fixes would include (but are not limited to) including a contributions section in the repo README or including a separate CONTRIBUTING.md file (e.g. [Contributor Covenant](https://www.contributor-covenant.org/)).

The contribution guidelines should guide "third parties wishing to 1) Contribute to the software 2) Report issues or problems with the software 3) Seek support".

Part of [this JOSS review](https://github.com/openjournals/joss-reviews/issues/3540).
