### 2021-08-01 #3540 @rrrlw
Minor (and optional) note: lines 33 and 36 in the article proof could be formatted as code with backticks (`) in markdown.
