### 2021-08-01 VBlesius/RHRT#5 @rrrlw
Warnings in automated tests
When running automated tests for the RHRT package on my local machine, I get 2 warnings (screenshots below). Are these significant issues?

*Screenshot for warning 1:*
![RHRT-TestWarning1](https://user-images.githubusercontent.com/6846515/127780385-9f4fb458-d65f-4897-998d-9017b6f4fc57.png)

*Screenshot for warning 2:*
![RHRT-TestWarning2](https://user-images.githubusercontent.com/6846515/127780391-3de6b72d-e3f0-4dd6-a6c6-663486dcd81c.png)

Part of [this JOSS review](https://github.com/openjournals/joss-reviews/issues/3540).
