### 2021-08-02 #3540 @brunaw
Dear @VBlesius, thank you for your submission to the JOSS, here are my comments/suggestions:

(Note: I'm not repeating the same comments made/issues created by the first reviewer, but I agree with what has already been pointed out)

## Package/repository
- The license needs to be added to the package folder and not just the full repository
- The README last example is not reproducible due to it using external data. Try to limit the examples to something the users can directly reproduce from your code
- Linking the vignette to the README would be nice/easier to refer to
- The vignette is quite extensive. I think it would be better to split it into a few smaller/more specific vignettes
- The `checkForHRT()` function only returns an object if an`if()` condition is met. I suggest adding a results/return in case this `if` condition is `FALSE`

- The authors use a `constants.R` file instead of just  defining default values for the parameters. Is there a reason for this?

## Paper

- The summary has two paragraphs and one final sentence, which could all be better integrated to make the writing flow better
    - Even if some information is removed from the summary it can be later given in the text

- The Statement of need section does not clearly states the need for the package, just briefly discusses HRT and that `RHRT` is the first package to do this sort of analysis (note this is a software submission, so it should also focus in the actual proposed software)

- There is no state of the field section/writing, which should be part of the paper, so I think there are missing references
