### 2021-08-10 64a9bf2
adds copyright and licence notices to each source file
