### 2021-08-10 2aeba4c
adds copyright and licence notices to each source file
