### 2021-08-10 VBlesius/RHRT#3 @VBlesius
Thank you for the feedback! The mentioned lines in the licence file serve as examples how to adapt ones program to the GPL. To our knowledge they are not meant to be adapted themselves.

We checked the [instructions](https://www.gnu.org/licenses/gpl-howto.html) of the licence again and found no mention of the need to adapt the licence file. However, we did not include copyright and licence notices which is recommended. We corrected this oversight here: 2aeba4cf710745fecd91b4a64b451cde8907b36c.
