### 2021-08-10 #3540 @VBlesius
> @VBlesius: the paper lists 2 authors, yourself and Andreas Dominik. I only see commits from you in the GitHub repository. Could you provide (brief) insight regarding the contributions of the second author?

@rrrlw Andreas Dominik is my PhD supervisor. Commits are only made by me because I implemented the package and wrote the paper, but essential implementation ideas, advice and feedback comes from Andreas Dominik. As I read in the [Author Guides](https://joss.readthedocs.io/en/latest/submitting.html#authorship) this is sufficient for co-authorship and therefore I included him as author.
