### 2021-08-10 316a022
adds passage about contributions and contributor covenant file
