### 2021-08-10 VBlesius/RHRT#4 @VBlesius
You are absolutely right, until now we did not mention contributions in the repo. I included a short passage inviting contributions in the README which I hope is sufficient (316a022deefd03687565a09a1a1fd85171874660).
And thank you very much for suggesting contributor covenant, I also included it in the repo.
