### 2021-08-10 #3540 @VBlesius
> Minor notes:
>     * line 24: CRAN has the square brackets, but lacks the following parentheses containing the URL to CRAN
>     * lines 33 and 36 in the article proof could be formatted as code with backticks (`) in markdown.

Thanks for the hints, fixed both :+1:
