### 2021-08-12 3ec2f6b
fixes #5: catches warning in tests
