### 2021-08-12 VBlesius/RHRT#5 @VBlesius
No, the warnings are important for users, but should not appear in the tests. Thank you for pointing this out! I fixed it here 3ec2f6ba2df4623ff3faecf2b78deaf01f77bb36.
