+++ RHRT/vignettes/objects_functions.Rmd
@@ -1,0 +1,138 @@
+<!---
+# Part of RHRT: R package to assess Heart Rate Turbulence from RR interval data 
+# Copyright (C) 2021 Valeria Blesius
+
+# RHRT is free software: you can redistribute it and/or modify
+# it under the terms of the GNU General Public License as published by
+# the Free Software Foundation, version 2 only.
+
+# RHRT is distributed in the hope that it will be useful,
+# but WITHOUT ANY WARRANTY; without even the implied warranty of
+# MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+# GNU General Public License for more details.
+
+# You should have received a copy of the GNU General Public License
+# along with RHRT.  If not, see <https://www.gnu.org/licenses/>.
+-->
+
+---
+title: "RHRT: Objects and Functions"
+author: "Valeria Blesius"
+date: "`r Sys.Date()`"
+output: 
+  rmarkdown::html_vignette:
+    keep_md: true
+vignette: >
+  %\VignetteIndexEntry{RHRT: Objects and Functions}
+  %\VignetteEngine{knitr::rmarkdown}
+  %\VignetteEncoding{UTF-8}
+---
+
+This vignette gives a detailed overview of all objects and functions provided by RHRT.
+
+--------
+
+## HRT Object
+
+### Slots
+
+An HRT object saves the data of one VPCS and its HRT results. It consists of the following slots:
+
+- **Intervals**
+  - `preRRs`: preceding regular intervals, 5 per default
+  - `couplRR`: CPI
+  - `compRR`: CMI
+  - `postRRs`: following regular intervals, 15 per default
+
+- **HRT Parameters**
+  - `TO`
+  - `TS`
+  - `TT`
+  - `nTS`: normalised TS, for more Information see chapter *Normalisation of Turbulence Slope* in [Scientific Background](background.html)
+
+- **Line coefficients**
+  - `intercept`: The intercept of the TS regression line that is needed for plotting
+  - `nintercept`: Analogously, the intercept of `nTS`
+
+### Functions
+
+- **`getRRs`** &nbsp;&nbsp; This function returns all intervals saved in the HRT object. This if helpful for i.e. calculating an averaged VPCS out of a set of HRT objects.
+
+- **`plot`** &nbsp;&nbsp; Per default `plot` displays a zoomed in plot of the VPCS with highlighted TO and TS and a legend given the rounded HRT parameter values. In addition to the common parameters of graphics.plot the method accepts the following parameters:
+  - `cropped`: switches between showing a zoomed in version and the full VPCS including CPI and CMI, the default is `TRUE`
+  - `add`: adds the plot to the current one
+  - `TT`: highlights TT and includes it to the legend
+  - `paramsLegend`: switches between showing and hiding the parameter values in the legend, the default is `TRUE`
+  - `colTO`, `colTS` and `colTT` determine the colour in which the different parameters are highlighted in the plot (red, blue and purple per default).
+
+## HRTList Object
+
+An HRTList object is created when `vectorToHRT` is called. All HRT parameters are calculated automatically in this process and can be called with the `getHRTParams`-methods.
+
+### Slots
+
+The HRTList object sums up all HRTs found in the given dataset. The slots are:
+
+- `name`: The name of the object if given to `vectorToHRT`
+- `IL`: the average length of all intervals in the given cleaned vector (see *vectorToHRT: Cleaning Input* for more information) 
+- `pos`: the indices of the CPIs as found in the given vector
+- `HRTs`: list of all HRT objects found
+- `avHRT`: an avHRT object averaged from all HRTs
+- `RMSSD`: the HRV parameter RMSSD calculated from all intervals in the given cleaned vector (see *vectorToHRT: Cleaning Input* for more information)
+
+### Functions
+
[…]
