### 2021-08-12 17938bc
corrects yaml position of vignettes and knits all
