+++ RHRT/vignettes/synopsis.md
@@ -1,7 +1,7 @@
 ---
 title: "RHRT: Synopsis for the hasty (Quick-start guide)"
 author: "Valeria Blesius"
-date: "`r Sys.Date()`"
+date: "2021-08-12"
 output: 
   rmarkdown::html_vignette:
     keep_md: true
@@ -19,7 +19,8 @@ This vignette sums up the most common functions and parameters needed when using
 
 ## Loading package and data
 
-```{r}
+
+```r
 library("RHRT")
 # testdataLong is a numeric vector of RR intervals in msec
 data("testdataLong", package = "RHRT")
@@ -33,13 +34,15 @@ ann <- testdataLong_Ann
 
 The **core function** of RHRT is `vectorToHRT` that finds valid VPCs in RR intervals and returns an `HRTList` object (see *HRTList object* in [Objects & Functions](objects_functions.html) for more information):
 
-```{r}
+
+```r
 hrtl <- vectorToHRT(ints) 
 ```
 
 Every RR interval sequence that matches the needed interval lengths is considered to be a coupling and compensatory interval of a VPC, which can lead to wrong matches. If your data is annotated, you can provide the **annotation data** with the parameters `annotations` and `PVCAnn`.
 
-```{r}
+
+```r
 hrtl <- vectorToHRT(ints, annotations = ann, PVCAnn = "V")
 ```
 
@@ -50,26 +53,47 @@ Other parameters are:
 * `normHallstrom` defines whether TS should be **normalised** with the method of Hallstrom et al. (see the chapter *Normalisation of Turbulence Slope* in the [scientific background](background.html) for more information). 
 
 ## Getting HRT parameters or class
-```{r}
+
+```r
 getResults(hrtl) # get the HRT class of the data
+```
+
+```
+## [1] "HRT0"
 ```
 
 Per default `getResults` checks whether all needed HRT parameters can be calculated reliably. This is done via a t-test per parameter value (for more information see chapter *Reliability Check* in the [scientific background](background.html) vignette). If any of the parameter values is **not reliable** `getResults` returns NR (not reliable). 
 
-```{r}
+
+```r
 getResults(hrtl, safe = FALSE) # get the HRT class without safety check
+```
+
+```
+## [1] "HRT0"
 ```
 
 In addition to the classification system HRT0-2 RHRT implements **HRTA-C** that is based on the three parameters TO, TS and TT. 
 
-```{r}
+
+```r
 getResults(hrtl, safe = FALSE, TT = TRUE) # include TT
+```
+
+```
+## [1] "HRTA"
 ```
 
 With the parameter `type` you can choose between getting only the HRT **class**, all **parameter values** or the parameter values with the corresponding **p-values** (types "class", "parameter" or "full", respectively).
 
-```{r}
+
+```r
 getResults(hrtl, type = "parameter", TT = TRUE) # get the averaged HRT parameters
+```
+
+```
+##        TO        TS        TT 
+## -10.57551  37.61417   3.00000
 ```
 
 Other parameters are:
@@ -80,15 +104,21 @@ Other parameters are:
 
 ## Plotting
 
-```{r, fig.width=7, fig.height=4}
+
+```r
 plot(hrtl, TT = TRUE) # plots the averaged VPCS and all underlying VPCSs in background
 ```
 
+![](synopsis_files/figure-html/unnamed-chunk-8-1.png)<!-- -->
+
 Per default the VPCS is **zoomed** in. If you want to also see the CPI and CMI use `cropped = FALSE`.
 
-```{r, fig.width=7, fig.height=4}
+
+```r
 plot(hrtl, cropped = FALSE) # shows also coupling and compensatory interval
 ```
+
+![](synopsis_files/figure-html/unnamed-chunk-9-1.png)<!-- -->
 
 --------
 
