+++ RHRT/vignettes/background.md
@@ -1,7 +1,7 @@
 ---
 title: "RHRT: Scientific Background"
 author: "Valeria Blesius"
-date: "`r Sys.Date()`"
+date: "2021-08-12"
 output: 
   rmarkdown::html_vignette:
     keep_md: true
