### 2021-08-12 c375bd6
adds links to vignettes in README
