+++ RHRT/README.md
@@ -72,4 +72,25 @@ chf201 <- read.table("~/some/path/chf201.csv")
 hrtl <- RHRT::vectorToHRT(chf201[[1]]*1000, ann = chf201[[2]])
 ```
 
-More example workflows can be found in the vignette.
+More example workflows can be found in the vignettes: [Quick start
+guide](vignettes/synopsis.md), [Objects &
+Functions](vignettes/objects_functions.md), [Scientific
+Background](vignettes/background.md) and [Example
+Pipelines](vignettes/examples.md).
+
+<!---
+# Part of RHRT: R package to assess Heart Rate Turbulence from RR interval data 
+# Copyright (C) 2021 Valeria Blesius
+
+# RHRT is free software: you can redistribute it and/or modify
+# it under the terms of the GNU General Public License as published by
+# the Free Software Foundation, version 2 only.
+
+# RHRT is distributed in the hope that it will be useful,
+# but WITHOUT ANY WARRANTY; without even the implied warranty of
+# MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+# GNU General Public License for more details.
+
+# You should have received a copy of the GNU General Public License
+# along with RHRT.  If not, see <https://www.gnu.org/licenses/>.
+-->
