### 2021-08-12 95489e2
changes links in vignettes from htmls to mds
