### 2021-08-12 15ace79
changes links from md to html or removes them for CRAN compatibility
