### 2021-08-12 94c073b
adds return value to checkForHRT
