### 2021-08-12 #3540 @VBlesius
Thank you very much for the thorough analysis of the package @brunaw!

> The license needs to be added to the package folder and not just the full repository

I added another licence file as suggested.

> The README last example is not reproducible due to it using external data. Try to limit the examples to something the users can directly reproduce from your code

Showing how to use external data is the point of this example, because this is the main field of application of the package. The example above uses only data included in the package and is therefore reproducible only from the given code as you suggested. The idea was to provide access to the package in different ways.

> Linking the vignette to the README would be nice/easier to refer to
> The vignette is quite extensive. I think it would be better to split it into a few smaller/more specific vignettes

As suggested I split the vignette into parts and added links between the vignette files and the README

> The checkForHRT() function only returns an object if anif() condition is met. I suggest adding a results/return in case this if condition is FALSE

If no HRT is found NULL is returned. However, it is true that this should be done explicitly. I added a return value in that case.

> The authors use a constants.R file instead of just defining default values for the parameters. Is there a reason for this?

Actually yes: The mentioned constants are default values used in several different functions. If one default value has to be adjusted, it just has to be changed in constants.R instead of in different functions.
