### 2021-08-13 211c114
smoothes summary
