+++ Paper/paper.bib
@@ -5,57 +5,57 @@
 	issn = {0140-6736},
 	url = {http://www.sciencedirect.com/science/article/pii/S0140673698084281},
 	doi = {10.1016/S0140-6736(98)08428-1},
-	abstract = {SummaryBackground
+	abstract = {{SummaryBackground}
 Identification of high-risk patients after acute myocardial infarction is essential for successful prophylactic therapy. The predictive accuracy of currently used risk predictors is modest even when several factors are combined. Thus, establishment of a new powerful method for risk prediction independent of the available stratifiers is of considerable practical value.
 Methods
-The study investigated fluctuations of sinus-rhythm cycle length after a single ventricular premature beat recorded in Holter electrocardiograms, and characterised the fluctuations (termed heart-rate turbulence) by two numerical parameters, termed turbulence onset and slope. The method was developed on a population of 100 patients with coronary heart disease and blindly applied to the population of the Multicentre Post-Infarction Program (MPIP; 577 survivors of acute infarction, 75 deaths during a median follow-up of 22 months) and to the placebo population of the European Myocardial Amiodarone Trial (EMIAT; 614 survivors of acute myocardial infarction, 87 deaths during median follow-up of 21 months). Multivariate risk stratification was done with the new parameters and conventional risk factors.
+The study investigated fluctuations of sinus-rhythm cycle length after a single ventricular premature beat recorded in Holter electrocardiograms, and characterised the fluctuations (termed heart-rate turbulence) by two numerical parameters, termed turbulence onset and slope. The method was developed on a population of 100 patients with coronary heart disease and blindly applied to the population of the Multicentre Post-Infarction Program ({MPIP}; 577 survivors of acute infarction, 75 deaths during a median follow-up of 22 months) and to the placebo population of the European Myocardial Amiodarone Trial ({EMIAT}; 614 survivors of acute myocardial infarction, 87 deaths during median follow-up of 21 months). Multivariate risk stratification was done with the new parameters and conventional risk factors.
 Findings
-One of the new parameters (turbulence slope) was the most powerful stratifier of follow-up mortality in EMIAT and the second most powerful stratifier in MPIP:MPIP risk ratio 3·5 (95\% Cl 2·2–5·5, p\&lt; 0·0001), EMIAT risk ratio 2·7 (1·8–4·2, p\&lt; 0·0001), In the multivariate analysis, low left-ventricular ejection fraction and turbulence slope were the only independent variables for mortality prediction in MPIP (p\&lt; 0·001), whereas in EMIAT, five variables were independent mortality predictors: abnormal turbulence onset, abnormal turbulence slope, history of previous infarction, low left-ventricular ejection fraction, and high mean heart rate (p\&lt; 0·001). In both MPIP and EMIAT, the combination of abnormal onset and slope was the most powerful multivariate risk stratifier: MPIP risk ratio 3·2 (1·7–6·0, p\&lt; 0·0001), EMIAT risk ratio 3·2 (1·8–5·6, p\&lt; 0·0001).
[…]
