### 2021-08-16 #3540 @VBlesius
Following, our changes made to the second part of @brunaw 's review regarding the paper:

>    The summary has two paragraphs and one final sentence, which could all be better integrated to make the writing flow better
>        Even if some information is removed from the summary it can be later given in the text

We tried to adapt the [instructions provided by JOSS](https://joss.readthedocs.io/en/latest/submitting.html#what-should-my-paper-contain) and restricted the paper to the sections summary, statement of need and example. Therefore general information and details are included in the summary. However, because it disrupted the writing flow as you pointed out, we incorporated the final sentence into the first paragraph.

>    The Statement of need section does not clearly states the need for the package, just briefly discusses HRT and that RHRT is the first package to do this sort of analysis (note this is a software submission, so it should also focus in the actual proposed software)
>    There is no state of the field section/writing, which should be part of the paper, so I think there are missing references

As suggested we revised the statement of need and added
- other HRT programs to give a glimpse of the state of HRT analysis tools
- a more detailed description of the scientific focus of RHRT and its corresponding key features
- example of the application of RHRT

I believe the paper improved through your comments and the changes made and I hope we could address the issues as expected.
