### 2021-09-17 33e02b1
adds description of RMSSD calculation to vignette
