### 2021-09-17 c8d86f2
adds more possibility to adjust or hide legend
