### 2021-09-17 821bc96
corrects examples in vignette
