+++ RHRT/vignettes/examples.Rmd
@@ -15,30 +15,31 @@ vignette: >
 
 ## Determining the HRT class of a person
 
-The main focus of the package is to determine the HRT parameters or class of a person by a long-term ECG measurement. Load the data as a numeric vector and use `vectorToHRT` to find HRTs, then `getResults` and `plot` to check the results:
+The main focus of the package is to determine the HRT parameters or class of a person by a long-term ECG measurement. Load the data as a numeric vector and use `vectorToHRT` to find HRTs, then `plot` and `getResults` to check the HRT:
 
 ```{r, fig.width=7, fig.height=4}
 library("RHRT")
 hrtl <- vectorToHRT(testdataLong) # create the HRTList
+plot(hrtl, main = "Zoomed in Tachogram") # plot the HRTs and check the variability
 getResults(hrtl) # get the averaged HRT parameters
-plot(hrtl) # plot the HRTs and check the variability
 ```
 
-The results do not pass the reliability check so we get "NR" as HRT class. The plot shows that firstly TO is near to zero and secondly there is a high variability in the VPCSs. We can go deeper into the data by checking the exact parameters (including TT as an additional hint to the person's status) and zooming out of the plot:
+The results do not pass the reliability check so we get "NR" instead of an HRT class. The plot shows that firstly TO is near to zero and secondly there is a high variability in the VPCSs. We can go deeper into the data by checking the exact parameters (including TT as an additional hint to the person's status) and zooming out of the plot:
 
 ```{r, fig.width=7, fig.height=4}
 round(
   getResults(hrtl, "full", TT = TRUE),
 digits = 2) # get the parameters and p-values of the variability check
-plot(hrtl, cropped = FALSE) # plot the full VPCSs
+plot(hrtl, cropped = FALSE, main = "Full Tachogram") # plot the full VPCSs
 ```
 
-As expected TO is not reliable with a p-value over 0.05. The VPCSs still seem to fluctuate a lot with the first two postRRs that determine TO varying above and below the ones of the avHRT. We can can get a picture on all TO values by using `getHRTParams`:
+As expected TO is not reliable with a p-value over 0.05. The VPCSs still seem to fluctuate a lot. We can can get a picture of the individual TO values by using `getHRTParams`:
 
 ```{r}
 tos <- getHRTParams(hrtl, "TO")
 tos
 summary(tos)
+par(mar=c(0, 3, 0, 0))
 boxplot(tos)
 ```
 
@@ -46,14 +47,14 @@ These results can help to come to a well-founded decision on whether to classify
 
 ## Comparing HRT results with different methodological parameters
 
-This is an example how the package can be used to analyse the HRT methodology. Due to data size we use a set of HRTs of one person rather than a set of averaged HRTs. For example, we can compare the difference in `TS` values when using different `numPostRRs`.
+This is an example how the package can be used to analyse the HRT methodology. For instance, we can compare the difference in `TO` and `TS` values when the order of the calculation steps are switched.
 
-```{r, fig.width=4, fig.height=4}
+```{r}
 library("RHRT")
-hrtl10 <- getHRTParams(vectorToHRT(testdataLong, numPostRRs = 10), "TS")
-hrtl20 <- getHRTParams(vectorToHRT(testdataLong, numPostRRs = 20), "TS")
-boxplot(hrtl10, hrtl20, names = c("TSRR = 10", "TSRR = 20"))
-t.test(hrtl10, hrtl20)
+hrtl <- vectorToHRT(testdataLong)
+getResults(hrtl, type = "parameter", safe = FALSE)
+hrtl@avHRT <- calcAvHRT(hrtl, orTO = "avBefore", orTS = "avAfter")
+getResults(hrtl, type = "parameter", safe = FALSE)
 ```
 
 --------
