+++ RHRT/vignettes/examples.md
@@ -1,7 +1,7 @@
 ---
 title: "RHRT: Example Pipelines"
 author: "Valeria Blesius"
-date: "2021-08-12"
+date: "2021-09-17"
 output: 
   rmarkdown::html_vignette:
     keep_md: true
@@ -15,26 +15,26 @@ vignette: >
 
 ## Determining the HRT class of a person
 
-The main focus of the package is to determine the HRT parameters or class of a person by a long-term ECG measurement. Load the data as a numeric vector and use `vectorToHRT` to find HRTs, then `getResults` and `plot` to check the results:
+The main focus of the package is to determine the HRT parameters or class of a person by a long-term ECG measurement. Load the data as a numeric vector and use `vectorToHRT` to find HRTs, then `plot` and `getResults` to check the HRT:
 
 
 ```r
 library("RHRT")
 hrtl <- vectorToHRT(testdataLong) # create the HRTList
-getResults(hrtl) # get the averaged HRT parameters
-```
-
-```
-## [1] "HRT0"
-```
-
-```r
-plot(hrtl) # plot the HRTs and check the variability
+plot(hrtl, main = "Zoomed in Tachogram") # plot the HRTs and check the variability
 ```
 
 ![](examples_files/figure-html/unnamed-chunk-1-1.png)<!-- -->
 
-The results do not pass the reliability check so we get "NR" as HRT class. The plot shows that firstly TO is near to zero and secondly there is a high variability in the VPCSs. We can go deeper into the data by checking the exact parameters (including TT as an additional hint to the person's status) and zooming out of the plot:
+```r
+getResults(hrtl) # get the averaged HRT parameters
+```
+
+```
+## [1] "NR"
+```
+
+The results do not pass the reliability check so we get "NR" instead of an HRT class. The plot shows that firstly TO is near to zero and secondly there is a high variability in the VPCSs. We can go deeper into the data by checking the exact parameters (including TT as an additional hint to the person's status) and zooming out of the plot:
 
 
 ```r
@@ -44,17 +44,17 @@ digits = 2) # get the parameters and p-values of the variability check
 ```
 
 ```
-##     TO     TS     TT    pTO    pTS    pTT 
-## -10.58  37.61   3.00   0.00   0.00   0.00
+##    TO    TS    TT   pTO   pTS   pTT 
+## -0.48 26.82  3.00  0.38  0.00  0.00
 ```
 
 ```r
-plot(hrtl, cropped = FALSE) # plot the full VPCSs
+plot(hrtl, cropped = FALSE, main = "Full Tachogram") # plot the full VPCSs
 ```
 
 ![](examples_files/figure-html/unnamed-chunk-2-1.png)<!-- -->
 
-As expected TO is not reliable with a p-value over 0.05. The VPCSs still seem to fluctuate a lot with the first two postRRs that determine TO varying above and below the ones of the avHRT. We can can get a picture on all TO values by using `getHRTParams`:
+As expected TO is not reliable with a p-value over 0.05. The VPCSs still seem to fluctuate a lot. We can can get a picture of the individual TO values by using `getHRTParams`:
 
 
 ```r
@@ -63,9 +63,9 @@ tos
 ```
 
 ```
-##  [1] -12.264829  -6.349883 -11.856998  -9.011983 -13.733660 -10.859878
-##  [7] -10.357064 -13.174994 -12.743754  -8.427299 -11.495815  -8.892588
-## [13] -11.795365  -6.502960 -11.165541
[…]
