+++ Changelog.md
@@ -5,8 +5,13 @@ All notable changes of RHRT will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 
-## [Unreleased] TBA
-
+## [1.1] 2021-09-17
+### Added
+- added parameters to the plot methods to hide and adjust legend
+### Changed
+- updated vignette and broke it down into smaller parts
+### Fixed
+- corrected the example pipelines in the vignette
 
 ## [1.0.1] 2021-06-28
 
