+++ RHRT/DESCRIPTION
@@ -1,8 +1,8 @@
 Package: RHRT
 Type: Package
 Title: Heart Rate Turbulence Analysis
-Version: 1.0.1
-Date: 2021-06-17
+Version: 1.1
+Date: 2021-09-17
 Authors@R: person(role = c("aut", "cre"), given = "Valeria", family = "Blesius", email = "valeria.blesius@mni.thm.de", comment = c(ORCID = "0000-0002-2391-242X"))
 BugReports: https://github.com/VBlesius/RHRT/issues
 Description: Methods to scan RR interval data for Premature Ventricular Complexes (PVCs) and parameterise and plot the resulting Heart Rate Turbulence (HRT). The methodology of HRT analysis is based on the original publication by Schmidt et al. <doi:10.1016/S0140-6736(98)08428-1> and extended with suggestions from <doi:10.1088/1361-6579/ab98b3>.
