### 2021-09-24 #3540 @brunaw
Dear @VBlesius apologies for such a big delay, I know it's very annoying to have your publication delayed because of this.

Thank you for the changes in the package, documentation & paper. I have been through your answers and it all seems good to me. One final thing, I didn't find the contribution guide (with how to 1) Contribute to the software 2) Report issues or problems with the software 3) Seek support), could you point me to where that is? I found the code of conduct, which is very good of you to have, but it's not the same thing

@bmcfee my review is almost done now, I just need this final answer from the authors. Apologies for the delay and thank you for your patience.
