### 2021-09-29 #3540 @VBlesius
Thank you for the final feedback @brunaw !
Community guidelines can be found in the readme of the repo [here](https://github.com/VBlesius/RHRT/blob/main/README.md).
