### 2021-10-14 4fdecbc
corrects and adds references
