+++ Paper/paper.bib
@@ -63,6 +63,7 @@ The absence of the heart rate turbulence after ventricular premature beats is a 
 	title = {Heart rate turbulence following ventricular premature beats in healthy controls},
 	volume = {8},
 	issn = {1082-720X},
+	doi = {10.1046/j.1542-474x.2003.08206.x},
 	abstract = {{BACKGROUND}: Heart rate turbulence ({HRT}) has recently been described as a strong, independent risk stratifier in postinfarct patients. To date, however, the incidence of false positive {HRT} findings in adults is unknown. Therefore, we performed a blinded, retrospective analysis of {HRT} in a prospectively collected database of 110 apparently healthy persons to determine the prevalence and clinical significance of abnormal {HRT} findings in healthy controls using previously published cut-off values.
 {METHODS} {AND} {RESULTS}: The study included 43 out of 110 apparently healthy adults, in whom a sufficient number of ventricular premature beats were available for {HRT} analysis on 24-hour Holter recordings. The {HRT} slope was dichotomized at 2.5 ms per R-R interval and {HRT} onset was dichotomized at 0\% using previously established criteria to define an abnormal {HRT} analysis in postinfarct patients. Using these definitions, abnormal {HRT} results were found in 2 out of 43 controls (5\%) for {HRT} slope and in 8 out of 43 controls (19\%) for {HRT} onset without any cardiac deaths during 32 +/- 15 months follow-up.
 {CONCLUSIONS}: The incidence of false positive {HRT} results in healthy middle-aged volunteers is low for {HRT} slope (5\%), but not for {HRT} onset (19\%) when previously published cut-off values are used to define abnormal {HRT} results. Thus, {HRT} slope dichotomized at 2.5 ms per R-R interval, but not {HRT} onset dichotomized at 0\%, may be used as a relatively specific tool for risk stratification in middle-aged persons.},
@@ -154,4 +155,11 @@ By offering the possibility to cross-validate the results of analyses obtained b
 	urldate = {2021-08-13},
 	date = {2016},
 	file = {Full Text PDF:/home/blesius/.mozilla/firefox/r1e11234.default/zotero/storage/H4DGDMYN/Pichot et al. - 2016 - HRVanalysis A Free Software for Analyzing Cardiac.pdf:application/pdf}
+}
+
+@article{blesius_comparability_2021,
+	title = {Comparability of Heart Rate Turbulence Methodology: 15 Intervals suffice to calculate Turbulence Slope – A methodological analysis using {PhysioNet} data of 1074 patients},
+	journaltitle = {[Manuscript submitted for publication]},
+	author = {Blesius, Valeria and Schölzel, Christopher and Ernst, Gernot and Dominik, Andreas},
+	date = {2021}
 }
