### 2021-10-14 #3540 @bmcfee
@VBlesius looks generally good to me!  I did notice that the license file in the repo is misspelled as "LICENCE" - that should probably be fixed, but I don't think it's worth generating a new version / archive for.
